+++ raytracing/tests/testsMatplotlib.py
@@ -1,5 +1,5 @@
 import unittest
-import env # modifies path
+import envtest # modifies path
 from raytracing import *
 import signal
 import time
